remove checkpoint change
