+++ contribs/src/test/scala/io/delta/storage/GCSLogStoreSuite.scala
@@ -16,8 +16,12 @@
 
 package io.delta.storage
 
-import org.apache.hadoop.fs.Path
-import org.apache.spark.sql.delta.{DeltaLog, FakeFileSystem, FakeGSCFileSystem, LogStoreSuiteBase}
+import java.net.URI
+
+import org.apache.hadoop.fs.{FSDataOutputStream, Path, RawLocalFileSystem}
+import org.apache.hadoop.fs.permission.FsPermission
+import org.apache.hadoop.util.Progressable
+import org.apache.spark.sql.delta.{FakeFileSystem, LogStoreSuiteBase}
 
 class GCSLogStoreSuite extends LogStoreSuiteBase {
 
@@ -40,5 +44,44 @@ class GCSLogStoreSuite extends LogStoreSuiteBase {
         store.write(new Path(s"gs://${tempDir.getCanonicalPath}", "1.json"), Iterator("foo"))
       }
     }
+  }
+}
+
+/**
+ * A fake GCS file system to verify delta commits are written in a separate gcs thread.
+ */
+class FakeGSCFileSystem extends RawLocalFileSystem {
+  override def getScheme: String = "gs"
+  override def getUri: URI = URI.create("gs:/")
+
+  private def assertGCSThread(f: Path): Unit = {
+    if (f.getName.contains(".json")) {
+      assert(
+        Thread.currentThread().getName.contains("delta-gcs-"),
+        s"writing $f was happening in non gcs thread: ${Thread.currentThread()}")
+    }
+  }
+
+  override def create(
+      f: Path,
+      permission: FsPermission,
+      overwrite: Boolean,
+      bufferSize: Int,
+      replication: Short,
+      blockSize: Long,
+      progress: Progressable): FSDataOutputStream = {
+    assertGCSThread(f)
+    super.create(f, permission, overwrite, bufferSize, replication, blockSize, progress)
+  }
+
+  override def create(
+      f: Path,
+      overwrite: Boolean,
+      bufferSize: Int,
+      replication: Short,
+      blockSize: Long,
+      progress: Progressable): FSDataOutputStream = {
+    assertGCSThread(f)
+    super.create(f, overwrite, bufferSize, replication, blockSize, progress)
   }
 }
